update to pypi packaging
