+++ protoc_polyglot/go/__init__.py
@@ -1,0 +1,1 @@
+from .cli import Lang_UI as cli
