+++ protoc_polyglot/java/__init__.py
@@ -1,0 +1,1 @@
+from .cli import Lang_UI as cli
