+++ protoc_polyglot/js/cli.py
@@ -1,20 +1,19 @@
 #!/usr/bin/python3
 import os, sys
 sys.path.insert(0, os.path.abspath('..'))
-from core.cli import *
+from protoc_polyglot.cli import *
 
 
 class Lang_UI(Base_UI):
     def _compile(self, dir_protos:str, dir_output: str, files:list[str]) -> None:
         shutil.rmtree(dir_output, ignore_errors=True)
-        os.makedirs(dir_output, exist_ok=True)
-        print(f'mkdir {dir_output}')
-
-        com = f'protoc -I {dir_protos} \
-        --plugin=protoc-gen-grpc_java={self.settings.plugin_path_java} \
-        --grpc_java_out={dir_output} \
-        --java_out={dir_output} \
-        {" ".join(files)}'
+        os.makedirs(dir_output, exist_ok=False)
+        
+        com = f"""protoc \
+        -I={dir_protos} \
+        {" ".join(files)} \
+        --js_out=import_style=commonjs:{dir_output} \
+        --grpc-web_out=import_style=commonjs,mode=grpcwebtext:{dir_output}"""
         
         print(com)
         os.system(com)
