+++ protoc_polyglot/python/cli.py
@@ -1,22 +1,22 @@
 #!/usr/bin/python3
 import os, sys
 sys.path.insert(0, os.path.abspath('..'))
-from core.cli import *
+from protoc_polyglot.cli import *
 
 
 class Lang_UI(Base_UI):
+    plugin_name = 'grpc_python_plugin'
+    
     def _compile(self, dir_protos:str, dir_output: str, files:list[str]) -> None:
         shutil.rmtree(dir_output, ignore_errors=True)
         os.makedirs(dir_output, exist_ok=False)
-        
-        com = f"""protoc \
-        -I={dir_protos} \
-        {" ".join(files)} \
-        --js_out=import_style=commonjs:{dir_output} \
-        --grpc-web_out=import_style=commonjs,mode=grpcwebtext:{dir_output}"""
+
+        com = f"""/usr/bin/protoc \
+        -I {dir_protos} \
+        --plugin=protoc-gen-grpc={self.get_plugin_path()} \
+        --grpc_out={dir_output} \
+        --python_out={dir_output} \
+        {" ".join(files)}"""
         
         print(com)
         os.system(com)
-
-if __name__ == '__main__':
-    Fire(Lang_UI)
