+++ upload-to-pypi.sh
@@ -9,3 +9,6 @@ hatch build
 twine upload -p $PYPI_KEY dist/*
 
 #ppackage/protobuf/bin/protoc -I ./samples --plugin=protoc-gen-grpc=ppackage/grpc/grpc_python_plugin --grpc_out=./output/python/bookclub --python_out=./output/python/bookclub bookclub/book.proto bookclub/member.proto bookclub/bookclub.proto
+
+pip install protoc-polyglot --upgrade
+pip install protoc-polyglot --upgrade
